Split dhis2 and gee entities, repositories and usecases
