+++ src/domain/dhis2/entities/DataElement.ts
