+++ src/domain/gee/entities/GeeData.ts
