+++ src/domain/utils.ts
